Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "This My Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp[speedups]>=3.12.15",
    "asyncio>=4.0.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "fastapi-jwt-auth>=0.5.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.35.0",
]
